+++ src/main/java/illyena/gilding/avengers/block/TeleportAnchorBlock.java
@@ -120,10 +120,10 @@ public class TeleportAnchorBlock extends BlockWithEntity implements BlockEntityP
                 int l = random.nextInt(2) * 2 - 1;
                 if (random.nextBoolean()) {
                     f = (double)pos.getZ() + 0.5 + 0.25 * (double)l;
-                    k = (double)(random.nextFloat() * 2.0F * (float)l);
+                    k = random.nextFloat() * 2.0F * (float)l;
                 } else {
                     d = (double)pos.getX() + 0.5 + 0.25 * (double)l;
-                    g = (double)(random.nextFloat() * 2.0F * (float)l);
+                    g = random.nextFloat() * 2.0F * (float)l;
                 }
 
                 world.addParticle(GildingParticles.STAR_PARTICLE, d, e, f, g, h, k);
